+++ tests/unit/test_main.py
@@ -678,9 +678,11 @@ class TestMain:
         assert response.status_code == 404
 
     # pylint: disable=too-many-statements
+    @patch('napps.kytos.topology.main.Main.notify_interface_status')
     @patch('napps.kytos.topology.main.Main.notify_topology_update')
-    async def test_enable_interfaces(self, mock_notify_topo):
+    async def test_enable_interfaces(self, *args):
         """Test enable_interfaces."""
+        (mock_notify_topo, mock_notify_interface) = args
         dpid = '00:00:00:00:00:00:00:01'
         mock_switch = get_switch_mock(dpid)
         mock_interface_1 = get_interface_mock('s1-eth1', 1, mock_switch)
@@ -717,6 +719,8 @@ class TestMain:
             interface_id
         )
         mock_notify_topo.assert_called()
+        assert mock_notify_interface.call_count == 1
+        assert mock_notify_interface.call_args[0][1] == 'enabled'
 
         mock_interface_1.enable.call_count = 0
         mock_interface_2.enable.call_count = 0
@@ -728,6 +732,8 @@ class TestMain:
         )
         assert mock_interface_1.enable.call_count == 1
         assert mock_interface_2.enable.call_count == 1
+        assert mock_notify_interface.call_count == 3
+        assert mock_notify_interface.call_args[0][1] == 'enabled'
 
         # test interface not found
         interface_id = '00:00:00:00:00:00:00:01:3'
@@ -747,10 +753,12 @@ class TestMain:
         assert mock_interface_1.enable.call_count == 0
         assert mock_interface_2.enable.call_count == 0
 
+    @patch('napps.kytos.topology.main.Main.notify_interface_status')
     @patch('napps.kytos.topology.main.Main.notify_link_enabled_state')
     @patch('napps.kytos.topology.main.Main.notify_topology_update')
-    async def test_disable_interfaces(self, mock_notify_topo, mock_noti_link):
+    async def test_disable_interfaces(self, *args):
         """Test disable_interfaces."""
+        (mock_notify_topo, mock_noti_link, mock_notify_interface) = args
         interface_id = '00:00:00:00:00:00:00:01:1'
         dpid = '00:00:00:00:00:00:00:01'
         mock_switch = get_switch_mock(dpid)
@@ -779,6 +787,8 @@ class TestMain:
         assert mock_noti_link.call_args[0][0] == mock_interface_1.link
         assert mock_noti_link.call_args[0][1] == "disabled"
         assert self.napp.topo_controller.disable_interface.call_count == 1
+        assert mock_notify_interface.call_count == 1
+        assert mock_notify_interface.call_args[0][1] == 'disabled'
         mock_notify_topo.assert_called()
 
         mock_interface_1.disable.call_count = 0
@@ -800,6 +810,8 @@ class TestMain:
         bulk_controller = self.napp.topo_controller.bulk_disable_links
         assert bulk_controller.call_count == 2
         assert len(bulk_controller.call_args[0][0]) == 1
+        assert mock_notify_interface.call_count == 3
+        assert mock_notify_interface.call_args[0][1] == 'disabled'
 
         # test interface not found
         interface_id = '00:00:00:00:00:00:00:01:3'
@@ -2179,3 +2191,18 @@ class TestMain:
 
         mock_detect_mismatched_link.return_value = False
         assert self.napp.link_status_mismatched(mock_link_1) is None
+
+    def test_notify_interface_status(self):
+        """Test notify_interface_status"""
+        mock_buffers_put = MagicMock()
+        self.napp.controller.buffers.app.put = mock_buffers_put
+        intf_mock = Mock()
+        expected_name = "kytos/topology.interface.disabled"
+        self.napp.notify_interface_status(intf_mock, 'disabled')
+        assert mock_buffers_put.call_count == 1
+        assert mock_buffers_put.call_args_list[0][0][0].name == expected_name
+
+        expected_name = "kytos/topology.interface.enabled"
+        self.napp.notify_interface_status(intf_mock, 'enabled')
+        assert mock_buffers_put.call_count == 2
+        assert mock_buffers_put.call_args_list[1][0][0].name == expected_name